Move CurrentSessionPollingStationId from domain to api::investigation
