+++ backend/src/api/investigation.rs
@@ -1,4 +1,8 @@
-use axum::{Json, extract::State, http::StatusCode};
+use axum::{
+    Json,
+    extract::{FromRef, FromRequestParts, Path, State},
+    http::{StatusCode, request::Parts},
+};
 use axum_extra::response::Attachment;
 use chrono::Datelike;
 use sqlx::{SqliteConnection, SqlitePool};
@@ -16,9 +20,8 @@ use crate::{
         data_entry::PollingStationResults,
         election::ElectionWithPoliticalGroups,
         investigation::{
-            CurrentSessionPollingStationId, PollingStationInvestigation,
-            PollingStationInvestigationConcludeRequest, PollingStationInvestigationCreateRequest,
-            PollingStationInvestigationUpdateRequest,
+            PollingStationInvestigation, PollingStationInvestigationConcludeRequest,
+            PollingStationInvestigationCreateRequest, PollingStationInvestigationUpdateRequest,
         },
         models::{ModelNa14_2Bijlage1Input, ToPdfFileModel},
         polling_station::{PollingStation, PollingStationId},
@@ -103,6 +106,33 @@ pub async fn delete_investigation_for_polling_station(
         }
     }
     Ok(())
+}
+
+pub struct CurrentSessionPollingStationId(pub PollingStationId);
+
+impl<S> FromRequestParts<S> for CurrentSessionPollingStationId
+where
+    SqlitePool: FromRef<S>,
+    S: Send + Sync,
+{
+    type Rejection = APIError;
+
+    async fn from_request_parts(parts: &mut Parts, state: &S) -> Result<Self, Self::Rejection> {
+        let path_extractor = Path::<PollingStationId>::from_request_parts(parts, state).await;
+        let pool = SqlitePool::from_ref(state);
+        let mut conn = pool.acquire().await?;
+
+        if let Ok(Path(id)) = path_extractor
+            && polling_station_repo::get(&mut conn, id).await.is_ok()
+        {
+            return Ok(CurrentSessionPollingStationId(id));
+        }
+
+        Err(APIError::NotFound(
+            "Polling station not found for the current committee session".to_string(),
+            ErrorReference::EntryNotFound,
+        ))
+    }
 }
 
 /// Create an investigation for a polling station
